clean up all updates and draws into one function
