+++ utils/utility.js
@@ -6,6 +6,7 @@ import {
   player,
   Enemies,
   secondsPassed,
+  FPS,
 } from "./config.js";
 import Coordinate from "./Coordinate.js";
 
@@ -14,4 +15,24 @@ function boardDraw() {
   context.fillRect(0, 0, BOARD.width, BOARD.height);
 }
 
-export { boardDraw,BOARD_HEIGHT,BOARD_WIDTH };
+function draw() {
+  boardDraw();
+  FPSDraw();
+  player.playerDraw();
+}
+
+function update() {
+  Enemies.forEach((enemy) => {
+    enemy.enemyUpdate();
+  });
+  player.playerUpdate();
+}
+
+function FPSDraw() {
+  // Draw FPS to the screen
+  context.font = "25px Arial";
+  context.fillStyle = "white";
+  context.fillText("FPS: " + FPS, 10, 30);
+}
+
+export { boardDraw, draw, update };
